+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for sub-account deletion via `account delete` command
+
 ## [3.3.0] - 2024-01-23
 
 ### Added